// packages/transform-ast/src/15/types.ts
export type Node = {
  ParseResult: ParseResult;
// ...
};
export interface ParseResult {
  version?: number;
  stmts?: RawStmt[];
}
// ...
export interface RawStmt {
  stmt?: Node;
  stmt_location?: number;
  stmt_len?: number;
}